Identity fix.
git-svn-id: http://svn.vuze.com/public/client@2788 c2c1b7af-1477-405b-86ee-8aefd1fc4dd2

+++ trunk/azureus2/src/org/gudy/azureus2/core3/peer/impl/transport/PEPeerTransportProtocol.java
@@ -161,6 +161,9 @@ PEPeerTransportProtocol
   private int writeSleepTime;
   private long lastReadTime;
   private long lastWriteTime;
+  
+  private boolean identityAdded = false;  //needed so we don't remove id's in closeAll() on duplicate connection attempts
+  
 	
 		// Protocol Info
 		
@@ -395,6 +398,7 @@ PEPeerTransportProtocol
   
 
   PeerIdentityManager.addIdentity( otherHash, otherPeerId, ip );
+  identityAdded = true;
  
   
 	//decode a client identification string from the given peerID
@@ -467,7 +471,7 @@ PEPeerTransportProtocol
 	}
 
   //remove identity
-  if ( this.id != null ) {
+  if ( this.id != null && identityAdded ) {
   	PeerIdentityManager.removeIdentity( manager.getHash(), this.id );
   }
